requirements-dataset.txt
numpy
pandas
scipy
psutil
joblib
tqdm
pathlib               
pyyaml
vtk
pyvista              
trimesh
meshio
pyigl                 
scikit-learn
matplotlib
plotly

torch                 
torch-geometric       
torch-scatter       
